refactor: move the tap binding alongside the runner instead of under it
