+++ knip.json
@@ -829,7 +829,7 @@
     "packages/app/src/prompt/prompt-app-types.ts": [
       "types"
     ],
-    "packages/app/src/runner/tap/commands.ts": [
+    "packages/app/src/tap/commands.ts": [
       "exports"
     ],
     "packages/app/src/studio/studio-app-types.ts": [
